+++ packages/toolboxes/evm/src/api/ethplorerApi.ts
@@ -1,43 +1,22 @@
-import {
-  assetFromString,
-  AssetValue,
-  filterAssets,
-  formatBigIntToSafeValue,
-  RequestClient,
-} from '@swapkit/helpers';
+import { formatBigIntToSafeValue, RequestClient } from '@swapkit/helpers';
 import { Chain } from '@swapkit/types';
 
 import type { AddressInfo } from '../types/ethplorer-api-types.ts';
 const baseUrl = 'https://api.ethplorer.io';
 
 export const ethplorerApi = (apiKey = 'freekey') => ({
-  getBalance: async (address: string, potentialScamFilter?: boolean) => {
-    const { getAddress } = await import('ethers');
+  getBalance: async (address: string) => {
     const { tokens = [] } = await RequestClient.get<AddressInfo>(
       `${baseUrl}/getAddressInfo/${address}`,
       { searchParams: { apiKey } },
     );
 
-    const balances = tokens.reduce((acc, token): AssetValue[] => {
-      const { symbol, decimals, address } = token.tokenInfo;
-
-      const tokenAsset = assetFromString(`${Chain.Ethereum}.${symbol}-${getAddress(address)}`);
-      if (!tokenAsset) return acc;
-
-      const balance = new AssetValue({
-        chain: tokenAsset.chain,
-        symbol: tokenAsset.symbol,
-        value: formatBigIntToSafeValue({
-          value: BigInt(token.rawBalance),
-          decimal: parseInt(decimals),
-        }),
-        decimal: parseInt(decimals),
-      });
-
-      return [...acc, balance];
-    }, [] as AssetValue[]);
-
-    return potentialScamFilter ? filterAssets(balances) : balances;
+    return tokens.map(({ tokenInfo: { symbol, decimals, address: tokenAddress }, rawBalance }) => ({
+      chain: Chain.Ethereum,
+      symbol: tokenAddress ? `${symbol}-${tokenAddress}` : symbol,
+      value: formatBigIntToSafeValue({ value: BigInt(rawBalance), decimal: parseInt(decimals) }),
+      decimal: parseInt(decimals),
+    }));
   },
 });