+++ packages/toolboxes/utxo/src/toolbox/BaseUTXOToolbox.ts
@@ -1,5 +1,5 @@
 import { HDKey } from '@scure/bip32';
-import { AssetValue, filterAssets, SwapKitNumber } from '@swapkit/helpers';
+import { AssetValue, SwapKitNumber } from '@swapkit/helpers';
 import type { UTXOChain } from '@swapkit/types';
 import { BaseDecimal, Chain, FeeOption } from '@swapkit/types';
 import { address as btcLibAddress, payments, Psbt } from 'bitcoinjs-lib';
@@ -334,11 +334,8 @@ export const BaseUTXOToolbox = (
     derivationPath: string;
   }) => (await createKeysForPath({ phrase, derivationPath, ...baseToolboxParams })).toWIF(),
 
-  getBalance: async (address: string, potentialScamFilter?: boolean) => {
-    const balances = await getBalance({ address, ...baseToolboxParams });
-
-    return potentialScamFilter ? filterAssets(balances) : balances;
-  },
+  getBalance: async (address: string, _potentialScamFilter?: boolean) =>
+    getBalance({ address, ...baseToolboxParams }),
 
   getFeeRates: () => getFeeRates(baseToolboxParams.apiClient),