+++ packages/wallets/evm-extensions/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @swapkit/wallet-evm-extensions
 
+## 1.0.0-rc.27
+
+### Patch Changes
+
+- perf improvements'
+
+- Updated dependencies []:
+  - @swapkit/toolbox-evm@1.0.0-rc.27
+
+## 1.0.0-rc.26
+
+### Patch Changes
+
+- Perf improvement on initialization of AssetValue
+
+- Updated dependencies []:
+  - @swapkit/toolbox-evm@1.0.0-rc.26
+
 ## 1.0.0-rc.25
 
 ### Patch Changes