+++ packages/wallets/wc/CHANGELOG.md
@@ -1,5 +1,25 @@
 # @swapkit/wallet-wc
 
+## 1.0.0-rc.29
+
+### Patch Changes
+
+- perf improvements'
+
+- Updated dependencies []:
+  - @swapkit/toolbox-cosmos@1.0.0-rc.27
+  - @swapkit/toolbox-evm@1.0.0-rc.27
+
+## 1.0.0-rc.28
+
+### Patch Changes
+
+- Perf improvement on initialization of AssetValue
+
+- Updated dependencies []:
+  - @swapkit/toolbox-cosmos@1.0.0-rc.26
+  - @swapkit/toolbox-evm@1.0.0-rc.26
+
 ## 1.0.0-rc.27
 
 ### Patch Changes